refactor: 숫자 생성 전략에 따라 로또를 생성하는 역할 Lotto가 수행하게 리팩터링
